# apps/backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Amber — Rare Disease Atlas API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/me/settings": {
      "patch": {
        "tags": [
          "account"
        ],
        "summary": "Update Settings",
        "description": "Role, language, expert mode, 16+ confirmation. First sign-in role choice uses this.",
        "operationId": "updateSettings",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionUser"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: lists invalid field names only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me/export": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Export My Data",
        "description": "All user data as JSON.",
        "operationId": "exportMyData",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataExport"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Delete Me",
        "description": "Delete the account and everything in it; clears the session.",
        "operationId": "deleteMe",
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/profile": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Get Profile",
        "description": "The user's PatientProfile.",
        "operationId": "getProfile",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfile"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "account"
        ],
        "summary": "Put Profile",
        "description": "Replace the PatientProfile.",
        "operationId": "putProfile",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatientProfile"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientProfile"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: lists invalid field names only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/consents": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "List Consents",
        "description": "The user's consents.",
        "operationId": "listConsents",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Consent"
                  },
                  "type": "array",
                  "title": "Response Listconsents"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "account"
        ],
        "summary": "Grant Consent",
        "description": "Grant a consent.",
        "operationId": "grantConsent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentGrant"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Consent"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: lists invalid field names only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/consents/{consent_type}": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Revoke Consent",
        "description": "Revoke a consent and delete the data held under it.",
        "operationId": "revokeConsent",
        "parameters": [
          {
            "name": "consent_type",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ConsentType"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/contributions": {
      "get": {
        "tags": [
          "contributions"
        ],
        "summary": "List Contributions",
        "description": "The user's own contributions.",
        "operationId": "listContributions",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Contribution"
                  },
                  "type": "array",
                  "title": "Response Listcontributions"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "contributions"
        ],
        "summary": "Create Contribution",
        "description": "Contribute a patient-reported profile or asset (status pending_review).",
        "operationId": "createContribution",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContributionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Contribution"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "consent_required (no active consent of the needed type) or forbidden.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: lists invalid field names only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/contributions/{contribution_id}": {
      "delete": {
        "tags": [
          "contributions"
        ],
        "summary": "Delete Contribution",
        "description": "Delete one of the user's contributions.",
        "operationId": "deleteContribution",
        "parameters": [
          {
            "name": "contribution_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Contribution Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/edges/{edge_id}/flag": {
      "post": {
        "tags": [
          "contributions"
        ],
        "summary": "Flag Edge",
        "description": "Flag an edge; it is shown as under_review.",
        "operationId": "flagEdge",
        "parameters": [
          {
            "name": "edge_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Edge Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FlagCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FlagResult"
                }
              }
            }
          },
          "401": {
            "description": "sign_in_required: guest, or the session expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "not_found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "validation_error: lists invalid field names only.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "501": {
            "description": "not_implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Consent": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "consent_type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "version": {
            "type": "string",
            "title": "Version",
            "description": "Version of the consent text the user agreed to."
          },
          "granted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Granted At"
          },
          "revoked_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revoked At"
          },
          "about_child": {
            "type": "boolean",
            "title": "About Child",
            "default": false
          },
          "parental_responsibility_confirmed": {
            "type": "boolean",
            "title": "Parental Responsibility Confirmed",
            "default": false
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "description": "True while not revoked."
          }
        },
        "type": "object",
        "required": [
          "id",
          "consent_type",
          "version",
          "granted_at",
          "active"
        ],
        "title": "Consent"
      },
      "ConsentGrant": {
        "properties": {
          "consent_type": {
            "$ref": "#/components/schemas/ConsentType"
          },
          "version": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Version",
            "description": "Consent text version shown."
          },
          "about_child": {
            "type": "boolean",
            "title": "About Child",
            "description": "Consent covers data about a child.",
            "default": false
          },
          "parental_responsibility_confirmed": {
            "type": "boolean",
            "title": "Parental Responsibility Confirmed",
            "description": "Required true when about_child is true.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "consent_type",
          "version"
        ],
        "title": "ConsentGrant"
      },
      "ConsentType": {
        "type": "string",
        "enum": [
          "upload",
          "contribute"
        ],
        "title": "ConsentType"
      },
...
      "DataExport": {
        "properties": {
          "exported_at": {
            "type": "string",
            "format": "date-time",
            "title": "Exported At"
          },
          "account": {
            "$ref": "#/components/schemas/AccountInfo"
          },
          "settings": {
            "$ref": "#/components/schemas/ProfileSettings"
          },
          "openai_connected": {
            "type": "boolean",
            "title": "Openai Connected",
            "description": "Tokens exist (the tokens are not exported)."
          },
          "consents": {
            "items": {
              "$ref": "#/components/schemas/Consent"
            },
            "type": "array",
            "title": "Consents"
          },
          "patient_profile": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PatientProfile"
              },
              {
                "type": "null"
              }
            ]
          },
          "chat_sessions": {
            "items": {
              "$ref": "#/components/schemas/ChatSessionExport"
            },
            "type": "array",
            "title": "Chat Sessions"
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/Document"
            },
            "type": "array",
            "title": "Documents"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "contributions": {
            "items": {
              "$ref": "#/components/schemas/Contribution"
            },
            "type": "array",
            "title": "Contributions"
          },
          "edge_flags": {
            "items": {
              "$ref": "#/components/schemas/EdgeFlag"
            },
            "type": "array",
            "title": "Edge Flags"
          },
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/Job"
            },
            "type": "array",
            "title": "Jobs"
          }
        },
        "type": "object",
        "required": [
          "exported_at",
          "account",
          "settings",
          "openai_connected",
          "consents",
          "patient_profile",
          "chat_sessions",
          "documents",
          "findings",
          "contributions",
          "edge_flags",
          "jobs"
        ],
        "title": "DataExport",
        "description": "Everything stored about the user (GDPR Art. 15/20, CCPA right to know)."
      },
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorDetail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "description": "The single error envelope used by every endpoint."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PatientProfile": {
        "properties": {
          "diseases": {
            "items": {
              "$ref": "#/components/schemas/ProfileDisease"
            },
            "type": "array",
            "title": "Diseases"
          },
          "genes": {
            "items": {
              "$ref": "#/components/schemas/ProfileGene"
            },
            "type": "array",
            "title": "Genes"
          },
          "variants": {
            "items": {
              "$ref": "#/components/schemas/ProfileVariant"
            },
            "type": "array",
            "title": "Variants"
          },
          "phenotypes": {
            "items": {
              "$ref": "#/components/schemas/ProfilePhenotype"
            },
            "type": "array",
            "title": "Phenotypes"
          },
          "age_years": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Years",
            "description": "Age in years (no birth date)."
          },
          "age_range": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgeRange"
              },
              {
                "type": "null"
              }
            ],
            "description": "Age range, alternative to age_years."
          },
          "onset": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Onset",
            "description": "Disease onset: HPO onset term ID or label."
          },
          "country": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Z]{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country",
            "description": "Optional ISO 3166-1 alpha-2 country, only to find nearby groups."
          },
          "about_child": {
            "type": "boolean",
            "title": "About Child",
            "description": "The profile describes a child.",
            "default": false
          },
          "parental_responsibility_confirmed": {
            "type": "boolean",
            "title": "Parental Responsibility Confirmed",
            "description": "Required true when about_child is true.",
            "default": false
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "Server-set last update time."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PatientProfile"
      },
...
      "SessionUser": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ],
            "description": "Null until the user picks a role."
          },
          "role_verified": {
            "type": "boolean",
            "title": "Role Verified",
            "default": false
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "expert_mode": {
            "type": "boolean",
            "title": "Expert Mode",
            "default": false
          },
          "age_confirmed": {
            "type": "boolean",
            "title": "Age Confirmed",
            "description": "User confirmed they are 16 or older.",
            "default": false
          },
          "consents": {
            "items": {
              "$ref": "#/components/schemas/ConsentType"
            },
            "type": "array",
            "title": "Consents",
            "description": "Active consents."
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "SessionUser"
      },
      "SettingsUpdate": {
        "properties": {
          "role": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Role"
              },
              {
                "type": "null"
              }
            ],
            "description": "patient, doctor or researcher (not guest)."
          },
          "language": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[a-z]{2,3}(-[A-Za-z0-9]{2,8})?$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "expert_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expert Mode"
          },
          "age_confirmed_16": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age Confirmed 16",
            "description": "Confirm the user is 16 or older."
          }
        },
        "type": "object",
        "title": "SettingsUpdate"
      },
...
    }
  }
}